backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
asyncpg
qdrant-client
google-generativeai
langchain
langchain-community
langchain-text-splitters
python-dotenv
pytest
httpx
alembic
psycopg2-binary
pyjwt
passlib
bcrypt
python-multipart
slowapi
prometheus-client
fastapi-utils
langchain-google-genai
langchain-google-genai[embeddings]